# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Campus Second-hand Trading Platform API
  version: 0.1.0
servers:
  - url: http://localhost:7001
paths:
  /api/auth/register:
    post:
      operationId: register
      summary: 用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: 注册成功，返回用户信息和认证凭证
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: 输入验证失败
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 用户名已存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      operationId: login
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: 登录成功，返回用户信息和认证凭证
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/logout:
    post:
      operationId: logout
      summary: 用户注销
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 注销成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    get:
      operationId: getCurrentUser
      summary: 获取当前登录用户信息
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 当前用户信息
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          description: 未登录或登录已过期
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/password:
    put:
      operationId: updatePassword
      summary: 修改密码
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordInput"
      responses:
        "200":
          description: 密码修改成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: 当前密码错误或新密码验证失败
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/cart:
    post:
      operationId: addToCart
      summary: 加入购物车（需认证）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartInput"
      responses:
        "201":
          description: 已加入购物车
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: 商品已下架或售罄
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listCart
      summary: 查看购物车（需认证）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 购物车商品列表
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CartItem"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cart/{id}:
    delete:
      operationId: removeFromCart
      summary: 删除购物车商品（需认证）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 已从购物车移除
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 购物车项不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cart/{id}/select:
    patch:
      operationId: toggleCartSelect
      summary: 切换购物车商品勾选状态（需认证）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [selected]
              properties:
                selected:
                  type: boolean
      responses:
        "200":
          description: 勾选状态已更新
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 购物车项不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cart/checkout:
    post:
      operationId: checkout
      summary: 结算勾选商品（需认证）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: 结算成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      orderId:
                        type: integer
        "400":
          description: 库存不足或支付密码错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: 余额不足
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cart/buy-now:
    post:
      operationId: buyNow
      summary: 立即购买（需认证）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyNowInput"
      responses:
        "200":
          description: 购买成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      orderId:
                        type: integer
        "400":
          description: 商品已售罄或支付密码错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: 余额不足
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: hex
  schemas:
    User:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
    RegisterInput:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          pattern: "^[一-龥a-zA-Z0-9]+$"
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    UpdatePasswordInput:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
    AuthResponse:
      type: object
      required: [data, token]
      properties:
        data:
          $ref: "#/components/schemas/User"
        token:
          type: string
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    AddToCartInput:
      type: object
      additionalProperties: false
      required: [itemId, quantity]
      properties:
        itemId:
          type: integer
          description: 商品 ID
        quantity:
          type: integer
          minimum: 1
          description: 数量
    CartItem:
      type: object
      required:
        [
          id,
          userId,
          itemId,
          quantity,
          selected,
          createdAt,
          title,
          price,
          coverImage,
          stock,
          status,
        ]
      properties:
        id:
          type: integer
        userId:
          type: integer
        itemId:
          type: integer
        quantity:
          type: integer
        selected:
          type: boolean
        createdAt:
          type: string
          format: date-time
        title:
          type: string
        price:
          type: number
        coverImage:
          type: string
        stock:
          type: integer
        status:
          type: string
    CheckoutInput:
      type: object
      additionalProperties: false
      required: [paymentPassword]
      properties:
        paymentPassword:
          type: string
          description: 支付密码
    BuyNowInput:
      type: object
      additionalProperties: false
      required: [itemId, quantity, paymentPassword]
      properties:
        itemId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        paymentPassword:
          type: string
...